+++ test/unit/core/mocks/FilesHandler.mocks.js
@@ -9,6 +9,7 @@ Unless required by applicable law or agreed to in writing, software distributed 
 */
 
 const sinon = require("sinon");
+const Behavior = require("../../../../src/mocks/Behavior");
 
 jest.mock("../../../../src/mocks/FilesHandler");
 
@@ -16,64 +17,32 @@ const FilesHandler = require("../../../../src/mocks/FilesHandler");
 
 const INITIAL_FILES = {
   file1: {
-    behavior1: {
-      fixtures: [
-        {
-          url: "/api/foo/foo-uri",
-          method: "GET",
-          response: {
-            status: 200,
-            body: {
-              fooProperty: "foo"
-            }
-          }
-        }
-      ],
-      totalFixtures: 1,
-      methods: {
-        POST: {
-          "/api/foo/foo-uri": {
-            route: "foo-route-parser",
-            response: {
-              status: 200,
-              body: {
-                fooProperty: "foo"
-              }
-            }
+    behavior1: new Behavior([
+      {
+        url: "/api/foo/foo-uri",
+        method: "GET",
+        response: {
+          status: 200,
+          body: {
+            fooProperty: "foo"
           }
         }
       }
-    }
+    ])
   },
   file2: {
-    behavior2: {
-      fixtures: [
-        {
-          url: "/api/foo/foo-uri-2",
-          method: "POST",
-          response: {
-            status: 422,
-            body: {
-              fooProperty2: "foo2"
-            }
-          }
-        }
-      ],
-      totalFixtures: 1,
-      methods: {
-        POST: {
-          "/api/foo/foo-uri-2": {
-            route: "foo-route-parser",
-            response: {
-              status: 422,
-              body: {
-                fooProperty2: "foo2"
-              }
-            }
+    behavior2: new Behavior([
+      {
+        url: "/api/foo/foo-uri-2",
+        method: "POST",
+        response: {
+          status: 422,
+          body: {
+            fooProperty2: "foo2"
           }
         }
       }
-    }
+    ])
   },
   folder: {
     folder2: {
@@ -84,9 +53,18 @@ const INITIAL_FILES = {
   }
 };
 
+INITIAL_FILES.file1.behavior1._lastPath = "behavior1";
+INITIAL_FILES.file2.behavior2._lastPath = "behavior2";
+
+const INITIAL_CONTENTS = [INITIAL_FILES.file1.behavior1, INITIAL_FILES.file2.behavior2, {}];
+
 class Mock {
   static get files() {
     return INITIAL_FILES;
+  }
+
+  static get contents() {
+    return INITIAL_CONTENTS;
   }
 
   constructor() {
@@ -94,6 +72,7 @@ class Mock {
 
     this._stubs = {
       files: INITIAL_FILES,
+      contents: INITIAL_CONTENTS,
       init: this._sandbox.stub().resolves(),
       start: this._sandbox.stub().resolves(),
       stop: this._sandbox.stub()
@@ -111,6 +90,7 @@ class Mock {
 
   restore() {
     this._stubs.files = INITIAL_FILES;
+    this._stubs.contents = INITIAL_CONTENTS;
     this._sandbox.restore();
   }
 }